WIP: Use newer GitHub action for Clippy check
This might later prove to be a mistake, given the close relationship
this action has with its predecessor, but this at least supports
reporting job failure when clippy fails.

WIP: permissions needed for results to show up when pr comes from a
local branch (otherwise, job fails, but it's not clear why - doesn't
happen in fork because fork is detected and results are printed to
stdout).