General cleanup of first enclave guide
This changes the info string of the fenced code blocks to more
accurately reflect the intention (console output versus a shell
script), uses the `--now` flag to eliminate some calls to `systemctl`,
uses `install` instead of doing the steps manually, and expands the
flag names for clarity. The file was also missing a newline at the
end.