Replace JMeter with Apache Bench in performance test
Apache Bench provides more useful output with far less configuration
out of the box than JMeter. The results are also more accurate since
the JMeter test was configured with a 20-second ramp-up, which
accounted for a large majority of the test and gave the impression
that there was no discernible speed-up when using more vCPUs.