Wait for ingress traffic before shutting down
Prior to this change, if the application within the enclave exited
shortly after sending a particularly slow (e.g. large) response, the
ingress proxy tasks would be aborted before the transfer could
complete, causing the response to be prematurely terminated. Instead
of aborting the tasks, send a shutdown signal and wait for their
completion.